HHH-6035 - Test classes merge
